requirements.txt
fastapi==0.105.0
SQLAlchemy==1.4.39
aiosqlite==0.19.0
pydantic==2.5.2
pytz==2022.1
dependency_injector==4.41.0
uvicorn==0.24.0
python-jose==3.3.0
redis==5.0.1
rq==1.15.1
scikit-learn==1.2.2
python-multipart==0.0.6
joblib==1.2.0
passlib==1.7.4
pyjwt==1.7.1
catboost==1.2
pandas==2.1.4
